+++ src/flickr_photos_api/api/license_methods.py
@@ -3,10 +3,11 @@ Methods for getting information about licenses from the Flickr API.
 """
 
 import functools
+import re
 
 from .base import FlickrApi
 from ..exceptions import LicenseNotFound
-from ..types import License
+from ..types import License, LicenseId
 
 
 class LicenseMethods(FlickrApi):
@@ -19,7 +20,10 @@ class LicenseMethods(FlickrApi):
     @functools.cache
     def get_licenses(self) -> dict[str, License]:
         """
-        Returns a list of licenses, arranged by code.
+        Returns a list of licenses, organised by numeric ID.
+
+        In particular, IDs can be looked up using the numeric ID
+        returned by many Flickr API methods.
 
         See https://www.flickr.com/services/api/flickr.photos.licenses.getInfo.htm
         """
@@ -29,7 +33,7 @@ class LicenseMethods(FlickrApi):
 
         # Add a short ID which can be used to more easily refer to this
         # license throughout the codebase.
-        license_ids = {
+        license_ids: dict[str, LicenseId] = {
             "All Rights Reserved": "all-rights-reserved",
             "Attribution-NonCommercial-ShareAlike License": "cc-by-nc-sa-2.0",
             "Attribution-NonCommercial License": "cc-by-nc-2.0",
@@ -65,22 +69,30 @@ class LicenseMethods(FlickrApi):
     @functools.cache
     def lookup_license_by_id(self, *, id: str) -> License:
         """
-        Given a license ID from the Flickr API, return the license data.
+        Return the license for a license ID.
 
-        e.g. a Flickr API response might include a photo in the following form:
+        The ID can be one of:
 
-                <photo license="0" …>
+        *   The numeric license ID returned from the Flickr API
+            (e.g. "0" ~> "All Rights Reserved")
+        *   The human-readable license ID returned from this library
+            (e.g. "cc-by-2.0" ~> "CC BY 2.0")
 
-        Then you'd call this function to find out what that means:
-
-                >>> api.lookup_license_by_id(id="0")
-                {"id": "all-rights-reserved", "name": "All Rights Reserved", "url": None}
-
-        See https://www.flickr.com/services/api/flickr.photos.licenses.getInfo.htm
         """
         licenses = self.get_licenses()
 
+        # If this is a numeric ID, then it must have come from the
+        # Flickr API.  Look it up directly in the dict.
+        if re.match(r"^[0-9]+$", id):
+            try:
+                return licenses[id]
+            except KeyError:
+                raise LicenseNotFound(license_id=id)
+
+        # Otherwise, this is a human-readable license ID from our
+        # library, so look for a matching license.
         try:
-            return licenses[id]
-        except KeyError:
+            matching_license = next(lic for lic in licenses.values() if lic["id"] == id)
+            return matching_license
+        except StopIteration:
             raise LicenseNotFound(license_id=id)